Remove app checkbox
